Add gate and custom signals
We've now added a gate with a CollissionPolygon2D and an Area2D for
reading when the player enteres the gate area. We also fire custom
signals when the player shoots their laser or throws their grenade.
Finally we also set the firing of lasers and launching of grenades to
only be available after a timeout.